Move apiproxy into a separate file
