+++ internal/jujuapi/streamproxy.go
@@ -6,11 +6,11 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/canonical/jimm/internal/auth"
-	"github.com/canonical/jimm/internal/dbmodel"
-	"github.com/canonical/jimm/internal/errors"
-	"github.com/canonical/jimm/internal/jimmhttp"
-	"github.com/canonical/jimm/internal/openfga"
+	"github.com/canonical/jimm/v3/internal/auth"
+	"github.com/canonical/jimm/v3/internal/dbmodel"
+	"github.com/canonical/jimm/v3/internal/errors"
+	"github.com/canonical/jimm/v3/internal/jimmhttp"
+	"github.com/canonical/jimm/v3/internal/openfga"
 	"github.com/gorilla/websocket"
 	"github.com/juju/juju/api/base"
 	jujuParams "github.com/juju/juju/rpc/params"
